% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allo}
\alias{allo}
\title{Allometric scaling data.}
\format{
A data frame with 455 rows.
\describe{
\item{Species}{}
\item{Mass}{}
\item{Temperature}{ }
\item{MR}{Metabolic rate}
\item{AvgMass}{}
\item{Q10SMR}{ }
}
}
\source{
\url{https://royalsocietypublishing.org/doi/suppl/10.1098/rsbl.2005.0378}
}
\usage{
allo
}
\description{
Used to introduce power laws.
}
\keyword{datasets}
