% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citymap}
\alias{citymap}
\title{Citymapper data.}
\format{
A data frame with 108 rows.
\describe{
\item{Date}{}
\item{Australia}{}
\item{Austria}{}
\item{Belgium}{}
\item{Brazil}{}
\item{Canada}{}
\item{Denmark}{}
\item{France}{}
\item{Germany}{}
\item{Italy}{}
\item{Japan}{}
\item{Mexico}{}
\item{Netherlands}{}
\item{Portugal}{}
\item{Russia}{}
\item{Singapore}{}
\item{South.Korea}{}
\item{Spain}{}
\item{Sweden}{}
\item{Turkey}{}
\item{United.Kingdom}{}
\item{United.States}{}
}
}
\source{
\url{https://citymapper.com/cmi/about}
}
\usage{
citymap
}
\description{
These are a bit unusual in that each country has a column.
}
\keyword{datasets}
