% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cntry}
\alias{cntry}
\title{Country data from Our World In Data.}
\format{
A data frame with 17,013 rows (current)
\describe{
\item{iso_code}{ISO 3-letter country code}
\item{location}{Text name of country}
\item{population}{}
\item{population_density}{}
\item{median_age}{}
\item{aged_65_older}{}
\item{aged_70_older}{}
\item{gdp_per_capita}{}
\item{extreme_poverty}{}
\item{cvd_death_rate}{}
\item{diabetes_prealence}{}
\item{female_smokers}{}
\item{male_smokers}{}
\item{handwashing_facilities}{}
\item{hospital_beds_per_100k}{}
\item{life_expectancy}{}
\item{alias}{Alias country name, shorter}
\item{lowstart}{Start of 'summer' viral respiratory low}
\item{lowend}{End of respiratory low. Sketchy at present.}
}
}
\source{
\url{https://github.com/owid/covid-19-data/tree/master/public/data}
and \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4847850/}
}
\usage{
cntry
}
\description{
Country data from Our World In Data.
}
\keyword{datasets}
