% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nanny.R
\name{corona}
\alias{corona}
\title{Basic setup of corona (Nanny Rona) R program}
\usage{
corona()
}
\description{
Try ?corona for help. For most functions, saying pdf=TRUE will write a PDF to images/.
If you wish to print to PDF, you need to setwd() to a directory that contains an
images/ directory that can be written to, or this will fail.
Individual examples are also available. Try e.g. ?corona_rabbits or ?corona_country
The results of corona_life() will depend on how your system handles animated GIF files.
}
\examples{
\donttest{
    corona_rabbits ( )
    corona_monty ( )
    corona_country ('France') 
    corona_vienna ( ) 
    corona_totals ( ) 
    country_dead ( ) 
    corona_converge ( )
    corona_metabolism ( ) 
    corona_citymap ( ) 
    corona_dowjones ( )
}

}
\keyword{Nanny}
\keyword{Rona}
\keyword{corona}
