% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nanny.R
\name{corona_all}
\alias{corona_all}
\title{Generate all Figures}
\usage{
corona_all()
}
\description{
For the book 'Rona' (printing to PDF)
work through and generate PDFs for all examples.
}
\keyword{Figures}
\keyword{Nanny}
\keyword{PDF}
\keyword{Rona}
\keyword{book}
\keyword{corona}
\keyword{print}
