% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_deaths.R
\name{corona_alldeaths}
\alias{corona_alldeaths}
\title{Plot UK deaths by week, with various adjustments:}
\usage{
corona_alldeaths(pdf = FALSE)
}
\arguments{
\item{pdf=FALSE}{will not print to PDF}
}
\description{
Assumes the existence of the data frame ukdead.
Draws three graphs:
\enumerate{
\item Raw data with a linear regression line;
\item Data with secular adjustment;
\item Data adjusted for a 'summer baseline' using the "other 9 years of data" after
secular adjustment.
}
}
\examples{
corona_alldeaths( pdf=TRUE ) 
}
\keyword{corona}
\keyword{rabbits}
