% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citymapper.R
\name{corona_citymap}
\alias{corona_citymap}
\title{Plot citymapper data against COVID-19 diagnoses, over time}
\usage{
corona_citymap(pdf = FALSE, FewCities = NULL, cols = 4)
}
\arguments{
\item{pdf}{= TRUE writes to PDF, default FALSE}

\item{cols}{Number of columns in output, default is 4}
}
\description{
Requires ggplot2, plyr and the data frames lock, owid, citymap.
Multiple, select frames are plotted.
}
\examples{
corona_citymap(pdf=TRUE, cols=4); 
}
\keyword{citymap}
\keyword{citymapper}
\keyword{corona}
\keyword{countries}
\keyword{daily}
\keyword{rates}
