% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{corona_country}
\alias{corona_country}
\title{Plot time course of coronavirus case incidence and deaths for one country}
\usage{
corona_country(country, pdf = FALSE, smooth = TRUE, deaths = TRUE)
}
\arguments{
\item{country}{: no default}

\item{pdf}{: defaults to FALSE. If TRUE, writes to country_name_new.pdf
i.e. '\emph{new.pdf' is appended to formal country name.
If the country name contains spaces ' ' they are changed to underscores '}'}

\item{smooth}{: default TRUE show smoothed (red) curve}

\item{deaths}{: default TRUE show deaths}
}
\description{
The daily case rate is also shown as a smoothed curve.
The smoothed death incidence is MULTIPLIED x5 to highlight
its relationship to the incidence curve. See grown-up documentation (LyX)
}
\examples{
corona_country('United States'); 
corona_country('Taiwan'); 
}
\keyword{corona}
\keyword{country}
\keyword{single}
