% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corona_data}
\alias{corona_data}
\title{The number of infections with the corresponsing country/ state and date}
\format{A data frame with  variables:
\describe{
  \item{state}{the state, e.g. US states like California or chines provinces such as Hubai}
  \item{country}{the country, e.g. Germany}
  \item{Lat}{Latitude, geodata for the country}
  \item{Long}{Longitude, geodata for the country}
  \item{date}{The day when the data was observed (infection, death, recoveries)}
  \item{infections}{The absolut number of infections in a country/state on the given date}
  \item{deaths}{The absolut number of deaths in a country/state on the given date}
  \item{recoveries}{The absolut number of recoveries in a country/state on the given date}
}}
\source{
\url{https://github.com/CSSEGISandData/COVID-19}
}
\usage{
corona_data
}
\description{
This dataset is downloaded from the John Hopkins University github account.
}
\keyword{datasets}
