% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowjones.R
\name{corona_dowjones}
\alias{corona_dowjones}
\title{Plot Dow-Jones Closing data.
Assumes the existence of the data frame djia, part of corona data.}
\usage{
corona_dowjones(pdf = FALSE)
}
\arguments{
\item{pdf=FALSE}{will not print to PDF}
}
\description{
Plot Dow-Jones Closing data.
Assumes the existence of the data frame djia, part of corona data.
}
\examples{
corona_dowjones ( pdf=TRUE ) 
}
\keyword{average}
\keyword{corona}
\keyword{dow}
\keyword{industrial}
\keyword{jones}
