% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{corona_monty}
\alias{corona_monty}
\title{A Monte Carlo simulation of the Monty Hall problem}
\usage{
corona_monty(runs = 100)
}
\arguments{
\item{runs}{specifies the number of parallel simulations, default=100.}
}
\description{
A Monte Carlo simulation of the Monty Hall problem
}
\examples{
corona_monty ( runs=10000 ) 
}
\keyword{Carlo}
\keyword{Hall}
\keyword{Monte}
\keyword{Monty}
\keyword{corona}
\keyword{coronavirus}
\keyword{simulation}
