% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbits.R
\name{corona_rabbits}
\alias{corona_rabbits}
\title{Demonstrate (graph) exponential growth of rabbit population:}
\usage{
corona_rabbits(topyear = 6, pdf = FALSE)
}
\arguments{
\item{topyear}{is last year, defaults to 6}

\item{pdf}{Will not print to PDF if FALSE (the default)}
}
\description{
For finer details, see the LyX/PDF documentation.
}
\examples{
corona_rabbits( topyear=10) 
}
\keyword{corona}
\keyword{rabbits}
