% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmelweis.R
\name{corona_vienna}
\alias{corona_vienna}
\title{Plot Semmelweis' original data from Vienna.}
\usage{
corona_vienna(pdf = FALSE)
}
\arguments{
\item{pdf}{default FALSE will \emph{not} print the two PDF files:
semmelweis_plot.pdf
semmelweis_run.pdf}
}
\description{
First simply 'plots the dots';
subsequently draws a run chart with a transition at the point where
he instituted hand-washing.
}
\examples{
\dontrun{
corona_vienna ( pdf=TRUE ) 
}
}
\keyword{Semmelweis}
\keyword{Vienna}
\keyword{corona}
