% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_deaths.R
\name{country_dead}
\alias{country_dead}
\title{Plot country deaths by week, with various adjustments:}
\usage{
country_dead(
  cc = "GBRTENW",
  pdf = FALSE,
  summerstart = 6,
  summerend = 8,
  save = FALSE
)
}
\arguments{
\item{summerstart, }{summerend (which months, e.g. 6=June)}

\item{save}{Do we save the data as a CSV}

\item{pdf=FALSE}{will not print to PDF}
}
\description{
Assumes the existence of the data frame ccdead where cc is one of AUT, BEL, BGR, CHE, CZE, DEUTNP, DNK, ESP,
EST, FIN, FRATNP, GBRTENW, GBR_SCO, HUN, ISL, ITA, LTU, LUX, NLD, NOR, PRT, SVK, SWE, or USA.
The unusual codes DEUTNP, FRATNP, GBRTENW, and GBR_SCO represent Germany, France, England+Wales and Scotland.
The default is GBRTENW, for reasons evident in the book.
}
\details{
The columns in the frame stmf are just 'PopCode', 'Year', 'Week', and 'Deaths'.

Draws three graphs:
\enumerate{
\item Raw data with a linear regression line, over n years;
\item Data with secular adjustment;
\item Data adjusted for a 'summer baseline' using the "other n years of data" after
secular adjustment.
}
}
\examples{
country_dead( pdf=TRUE ) 
country_dead( cc='NZL', summerstart=12, summerend=2 ) 
}
\keyword{corona}
\keyword{deaths}
