% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{create_day_sequence}
\alias{create_day_sequence}
\title{Add day column to dataframe (grouped for each country and state)}
\usage{
create_day_sequence(data, number_of_cases = 100)
}
\arguments{
\item{data}{A dataframe (suits with dataframe generated by \code{read_corona()})}

\item{number_of_cases}{Defaul value: 100, Integer, indicating the minimum number of cases (infections/ deaths/ recovered)}
}
\value{
A dataframe (with an additional day colum as incremental integers starting at 1)
}
\description{
Day column where day = 1 corresponds to the first date with at least n `number_of_cases`
}
\examples{
\dontrun{
data <- read_corona()
data \%>\% create_day_sequence()
}

}
