% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-data.R
\name{create_doubling}
\alias{create_doubling}
\title{Create the doubling rates}
\usage{
create_doubling(corona_data)
}
\arguments{
\item{corona_data}{A dataframe imported with `read_corona`}
}
\value{
A numeric vector (representing) the number of cases given that the doubling rate is d days)
}
\description{
Creates a sequence of the number events given that the
doubling rate is t dates
}
