% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{djia}
\alias{djia}
\title{Historical Dow Jones Industrial Average prices.}
\format{
A data frame with 110 rows (current)
\describe{
\item{Date}{Date of transaction---excludes weekends etc}
\item{Open}{Opening average}
\item{High}{Maximum over the day}
\item{Low}{Minimum}
\item{Close}{Closing price}
}
}
\source{
\url{https://www.wsj.com/market-data/quotes/index/DJIA/historical-prices}
}
\usage{
djia
}
\description{
Historical Dow Jones Industrial Average prices.
}
\keyword{datasets}
