% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{filter_min_days}
\alias{filter_min_days}
\title{Filter dataframe for at least n cases per group}
\usage{
filter_min_days(data, min_n_days = 5)
}
\arguments{
\item{data}{Dataframe (containing a day column)}

\item{min_n_days}{Integer How many values must be available at least for each group}
}
\value{
A dataframe
}
\description{
The function computes the number of values for each group (e.g. country) and filters out those with less than the specified number of values.
It is recommended to first group the dataframe for the stata the number of values have to be evaluated (e.g. at least 10 values in a country)
}
