% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Model to the Data}
\usage{
fit_model(country, model_type = "Exponential")
}
\arguments{
\item{country}{Country/Region Name}

\item{model_type}{Model Type}
}
\description{
Fit a Model to the Data
}
