% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jh_data.R
\name{get_confirmed_cases}
\alias{get_confirmed_cases}
\title{Download Latest Covid Dataset}
\usage{
get_confirmed_cases()
}
\value{
data.table of confirmed cases by country and date
}
\description{
Download the latest dataset from the John Hopkins website
}
