% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jh_data.R
\name{get_latest_country_data}
\alias{get_latest_country_data}
\title{Get latest dataset for a country}
\usage{
get_latest_country_data(country)
}
\arguments{
\item{country}{name of Country/Region}
}
\description{
Long form dataset showing the daily tallies of confirmed cases since the first case was reported in the country.
Also shows the latest value and the first case in that country. These columns are useful for pivoting data from multiple countries
}
