% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-data.R
\name{lm_corona}
\alias{lm_corona}
\title{Create a growth model}
\usage{
lm_corona(corona_data)
}
\arguments{
\item{corona_data}{A dataframe (tibble) from \code{read_corona()}}
}
\value{
A tibble containing country, r squared, intercept, slope, base- and growth-rate
}
\description{
Build a growth model using the number of cases,
deaths or recoveries as dependent variable and
days,e.g. date as independent variable
}
\details{
The Intercept of the fitted model expresses the logarithmic base rate (log10)
and the coefficient represents the logarithmic growth rate (log10)
}
