% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lock}
\alias{lock}
\title{Approximate dates of full lockdown in various countries.}
\format{
A data frame with 110 rows (current)
\describe{
\item{iso_code}{Country}
\item{Lockdown}{Date of lockdown YYYY-MM-DD}
\item{nature}{Text description: national | partial | advice | empty(none) }
}
}
\source{
Various data sources.
}
\usage{
lock
}
\description{
Approximate dates of full lockdown in various countries.
}
\keyword{datasets}
