% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-data.R
\name{merge_data}
\alias{merge_data}
\title{Merge Infection, Death and Recover datasets}
\usage{
merge_data()
}
\value{
A dataframe with columns infections, deaths, recoveries (plus the constant columns state, country, lat, long, and date)
}
\description{
Merge Infection, Death and Recover datasets
}
\examples{
\dontrun{
data <- merge_data()
head(data)
}
}
