% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owid}
\alias{owid}
\title{Wide-ranging data from Our World In Data. I only use a tiny part.}
\format{
A data frame with 17,013 rows (current)
\describe{
\item{iso_code}{ISO 3-letter country code}
\item{date}{Date for this row of data}
\item{total_cases}{total cases to date}
\item{new_cases}{new cases}
\item{total_deaths}{eponymous}
\item{new_deaths}{}
\item{total_tests}{Recorded tests in toto}
\item{new_tests}{Eponymous}
\item{tests_units}{}
\item{stringency_index}{How severe the lockdown was}
}
}
\source{
\url{https://github.com/owid/covid-19-data/tree/master/public/data}
}
\usage{
owid
}
\description{
Wide-ranging data from Our World In Data. I only use a tiny part.
}
\keyword{datasets}
