% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{pivot_longer_by_date}
\alias{pivot_longer_by_date}
\title{Pivots date column names into one colum}
\usage{
pivot_longer_by_date(data, date_format = "\\\\d+\\\\/\\\\d+\\\\/\\\\d+")
}
\arguments{
\item{data}{A dataframe with data columns}

\item{date_format}{A regular expression to detect the date columns (default value matches DD/MM/YYY); works with dplyr's 'matches' function to select date columns)}
}
\value{
A dataframe in a long format (containing a new date column)
}
\description{
Pivots date column names into one colum
}
