% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_country}
\alias{plot_country}
\title{Plot the development in a single country}
\usage{
plot_country(data, show_model = TRUE)
}
\arguments{
\item{data}{Preprocessed data containing data for one country}

\item{show_model}{Logical, Should the predicted values be displayed}
}
\value{
A ggplot2 obj
}
\description{
Plot the development in a single country
}
\examples{
\dontrun{
data \%>\%
 preprocess_corona_data(statistic = "deaths",
                        countries = "Italy",
                        n = 10) \%>\%
 predict_growth() \%>\%
 plot_country(show_model = TRUE) +
 ggplot2::ggtitle("Corona  Death Growth Curve in Italy",
                  subtitle = "Starte date is the first day with > 10 deaths")
}
}
