% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot the growth curve}
\usage{
plot_growth(data)
}
\arguments{
\item{data}{A dataframe (or tibble) imported using \code{read_corona()}}
}
\value{
A ggplot object
}
\description{
Plot the growth curve
}
\examples{
\dontrun{
data(corona_data, statstisitic = "infections")
plot_growth(corona_data)
}
}
