% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{prediction}
\alias{prediction}
\title{Predict based on fitted model}
\usage{
prediction(model, duration)
}
\arguments{
\item{model}{Model Object returned from fit_model}

\item{duration}{Duration to Predict (from first case in country)}
}
\value{
Predictived Value
}
\description{
Make a prediction based on a fitted model and duration since first case in this country
}
