% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{preprocess_corona_data}
\alias{preprocess_corona_data}
\title{Preprocess the corona data for analysis}
\usage{
preprocess_corona_data(corona_data, statistic, countries, n)
}
\arguments{
\item{corona_data}{A dataframe (tibble) imported with `read_corona`}

\item{statistic}{A character string ("infections", "deaths" or "recovered")}

\item{countries}{A character vector containing the country names to include}

\item{n}{Integer, number of cases per day to include}
}
\value{
A dataframe (tibble)
}
\description{
Preprocess the corona data for analysis
}
\examples{
\dontrun{
data <- read_corona()
data \%>\%
 preprocess_corona_data(
   statistic = "infections", # Focus on infections
   countries = "Italy", # Focus on Italy
   n = 100) # Include days where the death toll exceeded 100 cases
 }


}
