% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_corona}
\alias{read_corona}
\title{Read the corona data}
\usage{
read_corona()
}
\value{
A tibble with
}
\description{
Returns a dataframe (tibble) with state, country, lat, long, date and poll numbers (infections, deaths and recoveries).
The returned dataframe is a unified version of three distinct datasets, each containing infection, deaths, recover data
}
\examples{
\dontrun{
read_corona()
}
}
