% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_deaths}
\alias{read_deaths}
\title{Read Corona Death Numbers}
\usage{
read_deaths(file = NULL)
}
\arguments{
\item{file}{The file path to load from. Do not specify it, to load from
JHU github repository (which is recommended)}
}
\value{
A tibble containing country, state, date, number of recoveries
}
\description{
Read Corona Death Numbers
}
\examples{
\dontrun{
read_deaths()
}
}
