% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_infections}
\alias{read_infections}
\title{Read Corona Infection Numbers}
\usage{
read_infections(file = NULL)
}
\arguments{
\item{file}{The file path to load from. Do not specify it, to load from
JHU github repository (which is recommended)}
}
\value{
A tibble containing country, state, date, number of infections
}
\description{
Read Corona Infection Numbers
}
\examples{
\dontrun{
read_infections()
}
}
