% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_raw_corona}
\alias{read_raw_corona}
\title{Reads a csv with JHU corona data}
\usage{
read_raw_corona(file)
}
\arguments{
\item{file}{The file path to data (in the main function it refers to the JHU github site)}
}
\value{
A tibble containing the JHU corona data
}
\description{
This function is mainly a helper function to read more
specific datasets (infections, deaths, recoveries)
}
\examples{
\dontrun{
file <- "URL"
read_corona_data(file)
}

}
