% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_recoveries}
\alias{read_recoveries}
\title{Read Corona Recovered Numbers}
\usage{
read_recoveries(file = NULL)
}
\arguments{
\item{file}{The file path to load from. Do not specify it, to load from
JHU github repository (which is recommended)}
}
\value{
A tibble containing country, state, date, number of recoveries
}
\description{
Read Corona Recovered Numbers
}
\examples{
\dontrun{
read_recoveries()
}
}
