% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{rename_corona_data}
\alias{rename_corona_data}
\title{Rename state and country column}
\usage{
rename_corona_data(data)
}
\arguments{
\item{data}{A dataframe with columnames `Country/Region` and `Province/State`}
}
\value{
A long dataframe (with two renamed colums)
}
\description{
Replace clumsy column names with accessable names
}
