% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{specify_countries}
\alias{specify_countries}
\title{Specify the countries to include}
\usage{
specify_countries(corona_data, countries)
}
\arguments{
\item{corona_data}{A dataframe (tibble) imported with `read_corona`}

\item{countries}{A character vector containing the country names to include}
}
\value{
A data.frame (tibble) with the data of inclueded countries
}
\description{
Specify the countries to include
}
\examples{
\dontrun{
data <- read_corona()
data \%>\%
 specify_countries(c("Italy", "Germany"))
}
}
