% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-data.R
\name{specify_statistic}
\alias{specify_statistic}
\title{Specify the statistic to analyse}
\usage{
specify_statistic(corona_data, statistic)
}
\arguments{
\item{corona_data}{A dataframe (tibble) imported with `read_corona`}

\item{statistic}{A character string}
}
\value{
A data.frame (tibble) with columns country, date, statistic
}
\description{
This function species the statistic to use in a
subsequent analysis. At the moment, three stistics are available:
}
\details{
1. infections
2. deaths
3. recoveries
}
\examples{
\dontrun{
data <- read_corona()
data \%>\%
 specify_statistic("infections")
}

}
