% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stmf}
\alias{stmf}
\title{Deaths, by week, for various countries.}
\format{
A data frame with 22678 rows.
\describe{
\item{iso_code}{Normally a 3-character country code e.g. NZL, AUT. England+Wales=GBRTENW, Scotland=GBR_SCO}
\item{Year}{YYYY}
\item{Week}{Week within that year, 1=1st}
\item{Deaths}{Number of deaths in that week}
\item{X}{}
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/weeklyprovisionalfiguresondeathsregisteredinenglandandwales} \url{https://www.stats.govt.nz/experimental/covid-19-data-portal}  \url{https://www.scb.se/en/finding-statistics/statistics-by-subject-area/population/population-composition/population-statistics/#_Tablesandgraphs} and also (registration now required) \url{https://www.mortality.org/}
}
\usage{
stmf
}
\description{
Deaths, by week, for various countries.
}
\keyword{datasets}
