% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jh_data.R
\name{tidy_confirmed_cases}
\alias{tidy_confirmed_cases}
\title{Tidy the Raw John Hopkins Data}
\usage{
tidy_confirmed_cases(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{
tidy dataset in long format
}
\description{
Tidy the Raw Data and Add the latest Stats for easy filtering
}
