% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukdead}
\alias{ukdead}
\title{UK Deaths, by week.}
\format{
A data frame with 537 rows.
\describe{
\item{Week}{A number starting at 1}
\item{Deaths}{Number of deaths in that week}
\item{Date}{The start of the week YYYY-MM-DD}
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/weeklyprovisionalfiguresondeathsregisteredinenglandandwales}
}
\usage{
ukdead
}
\description{
UK Deaths, by week.
}
\keyword{datasets}
