% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vienna}
\alias{vienna}
\title{Semmelweis' data on Deaths of parturients in Vienna}
\format{
A data frame with 98 rows
\describe{
\item{date}{Date of the start of each month YYYY-MM-01}
\item{births}{Number of births during that month}
\item{deaths}{Number of maternal deaths during that month}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Historical_mortality_rates_of_puerperal_fever}
}
\usage{
vienna
}
\description{
Semmelweis' data on Deaths of parturients in Vienna
}
\keyword{datasets}
