% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{visualise_fitted_vs_actual}
\alias{visualise_fitted_vs_actual}
\title{Visualise Fitted vs Actual Values}
\usage{
visualise_fitted_vs_actual(country, country_name = "",
  cutoff = Sys.Date())
}
\arguments{
\item{country}{Name of country to get data}

\item{country_name}{Used if you wish to specify a different name for The title in the Graph}

\item{cutoff}{If you wish to filter the data by a cutoff date enter a date here.}
}
\description{
Graph the trend in numbers and visually compare a fitted curve to the actual data projecting out for 3 days
}
