% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citymap}
\alias{citymap}
\title{Citymapper data.}
\format{
A data frame with 108 rows.
\describe{
\item{Date}{A column of dates in format YYYY-MM-DD}
\item{Country column names}{ Multiple country names, currently
"Australia"      "Austria"  "Belgium"  "Brazil"
"Canada"         "Date"     "Denmark"  "France"
"Germany"        "Italy"    "Japan"    "Mexico"
"Netherlands"    "Portugal" "Russia"   "Singapore"
"South.Korea"    "Spain"    "Sweden"   "Turkey"
"United.Kingdom" "United.States" }
}
\source{
\url{https://citymapper.com/cmi/about}
}
\usage{
citymap
}
\description{
These are a bit unusual in that each country has a column.
}
\keyword{datasets}
