% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nanny.R
\name{corona}
\alias{corona}
\title{Basic setup of corona (Nanny Covid) R program:
On loading corona loads ggplot2, gridExtra, plyr, qicharts
Use corona(dir='foo') to set the working directory to 'foo'
or just use setwd() as usual.}
\usage{
corona(dir = "C:/nc")
}
\arguments{
\item{dir}{'C:/nc' is default directory, a nasty Windows hack :(}
}
\description{
Basic setup of corona (Nanny Covid) R program:
On loading corona loads ggplot2, gridExtra, plyr, qicharts
Use corona(dir='foo') to set the working directory to 'foo'
or just use setwd() as usual.
}
\examples{
corona( dir='C:/foo/bar' ) 

You can then invoke the built-in functions like:
    corona_rabbits ( )
    
    corona_country ('France') 
    corona_dowjones ( )
    corona_lockdown ( ) 
    corona_vienna ( ) 
    corona_totals ( ) 
    corona_alldeaths ( ) 
    corona_citymap ( ) 

Each of these has associated help e.g. ?corona_vienna 
}
\keyword{Covid}
\keyword{Nanny}
\keyword{corona}
