% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nanny.R
\name{corona}
\alias{corona}
\title{Basic setup of corona (Nanny Rona) R program}
\usage{
corona(dir = "C:/nc")
}
\arguments{
\item{dir}{'C:/nc' is default MS Windows directory; on Linux or Mac, is '~/nc'}
}
\description{
Try ?corona for help. For most functions, saying pdf=TRUE will write PDF to images/.
On loading, corona also loads gganimate, ggplot2, gridExtra, plyr, qicharts2, reshape2
Use corona(dir='foo') to set the working directory to 'foo'
or just use setwd() as usual.
}
\examples{
\dontrun{
    corona_rabbits ( )
    corona_monty ( )
    corona_country ('France') 
    corona_lockdown ( cols=14 ) 
    corona_vienna ( ) 
    corona_totals ( ) 
    country_dead ( ) 
    corona_converge ( )
    corona_metabolism ( ) 
    corona_life ( ) 
    corona_citymap ( ) 
    corona_dowjones ( )
}

}
\keyword{Nanny}
\keyword{Rona}
\keyword{corona}
