% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nanny.R
\name{corona_all}
\alias{corona_all}
\title{Generate all Figures in the book 'I Nanny Rona' (printing to PDF)
Work through and generate PDFs for all examples.}
\usage{
corona_all()
}
\description{
Generate all Figures in the book 'I Nanny Rona' (printing to PDF)
Work through and generate PDFs for all examples.
}
\keyword{Covid}
\keyword{Figures}
\keyword{Nanny}
\keyword{PDF}
\keyword{book}
\keyword{corona}
\keyword{print}
