% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conway.R
\name{corona_life}
\alias{corona_life}
\title{Animate Conway's Game of Life
Depends on gganimate, reshape2}
\usage{
corona_life(filename = "default.gif", side = 50, steps = 100, density = 0.3)
}
\arguments{
\item{side}{The number of elements on the area's side (width or height)}

\item{steps}{The number of frames}

\item{density}{0.0--1 The density of the initial, random items}

\item{filename=foo.gif}{writes to this file name (NULL for current GIF device)}
}
\description{
Animate Conway's Game of Life
Depends on gganimate, reshape2
}
\examples{
corona_life ( filename='animation.gif', side=50, steps=500, density=0.2 ) 
}
\keyword{animation}
\keyword{conway}
\keyword{corona}
\keyword{frames}
\keyword{game}
\keyword{life}
\keyword{of}
