% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockdown.R
\name{corona_lockdown}
\alias{corona_lockdown}
\title{Draw multiple smoothed graphs of new daily cases, with lockdown date, if present}
\usage{
corona_lockdown(pdf = FALSE, minpeople = 4e+06, mincases = 200, cols = 5)
}
\arguments{
\item{pdf=FALSE}{By default will not print to PDF}
}
\description{
This requires the data frames lock and owid, as well as ggplot and plyr (ddply) libraries.
At present limited to countries with population > 4M,
and over 200 cases.
}
\examples{
corona_lockdown( cols=14 ) 
corona_lockdown( pdf=TRUE ) 
}
\keyword{corona}
\keyword{lockdown}
\keyword{smoothed}
