% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{corona_metabolism}
\alias{corona_metabolism}
\title{Allometric scaling of metabolic rates}
\usage{
corona_metabolism(pdf = FALSE, base = 10)
}
\arguments{
\item{pdf=FALSE}{will not print to PDF}
}
\description{
Log-log plot of mammalian weights (grams) against metabolic rates. The PDF file is allometry.pdf.
}
\examples{
corona_metabolism ( pdf=TRUE ) 
}
\keyword{Kleiber}
\keyword{allometric}
\keyword{allometry}
\keyword{corona}
\keyword{exponent}
\keyword{quarters}
\keyword{scaling}
\keyword{thirds}
\keyword{three}
\keyword{two}
