% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googletrends.R
\name{corona_trends}
\alias{corona_trends}
\title{Plot Google Trends data for searches involving the word 'coronavirus'.}
\usage{
corona_trends(pdf = FALSE)
}
\arguments{
\item{pdf}{default FALSE will \emph{not} print the PDF file}
}
\description{
Just plot the lines.
}
\examples{
corona_trends ( ) 
}
\keyword{Google}
\keyword{Trends}
\keyword{corona}
\keyword{coronavirus}
