% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googletrends.R
\name{corona_trends}
\alias{corona_trends}
\title{Plot Google Trends data for searches involving the word 'coronavirus'.
Just plot the lines.
Requires the CSV GoogleTrendsWorldCoronavirus.csv as the object gt}
\usage{
corona_trends(pdf = FALSE)
}
\arguments{
\item{pdf=FALSE}{will \emph{not} print thePDF file}
}
\description{
Plot Google Trends data for searches involving the word 'coronavirus'.
Just plot the lines.
Requires the CSV GoogleTrendsWorldCoronavirus.csv as the object gt
}
\examples{
corona_trends ( pdf=TRUE ) 
}
\keyword{Google}
\keyword{Trends}
\keyword{corona}
\keyword{coronavirus}
