% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gt}
\alias{gt}
\title{Google trends search for 'coronavirus'.}
\format{
A data frame with 155 rows (current)
\describe{
\item{Date}{Date in format YYYY-MM-DD}
\item{Day}{}
\item{coronavirus}{Coronavirus 'interest' as percentage of maximum count }
}
}
\source{
\url{https://trends.google.com/trends/}
}
\usage{
gt
}
\description{
Google trends search for 'coronavirus'.
}
\keyword{datasets}
