% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life}
\alias{life}
\title{The game of life.}
\format{
A data frame with 213 rows.
\describe{
\item{x}{x co-ordinate of an active cell}
\item{y}{y co-ordinate}
\item{pattern}{A name like 'blinker' --- will be common to several rows, specifying a Game of Life pattern }
}
}
\source{
(internal generation)
}
\usage{
life
}
\description{
This specifies initial conditions, using a clumsy storage format as below.
}
\keyword{datasets}
