% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_infections}
\alias{read_infections}
\title{Read Corona Infection Numbers}
\usage{
read_infections()
}
\value{
A tibble containing country, state, date, number of infections
}
\description{
Read Corona Infection Numbers
}
\examples{
\dontrun{
read_infections()
}
}
