% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_raw_corona}
\alias{read_raw_corona}
\title{Reads a csv with JHU corona data}
\usage{
read_raw_corona(file)
}
\arguments{
\item{file}{}
}
\value{
A tibble containing the JHU corona data
}
\description{
This function is mainly a helper function to read more specific datasets (infections, deaths, recoveries)
}
\examples{
\dontrun{
file <- "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_19-covid-Deaths.csv" # '
read_corona_data(file)
}

}
